+++ core-ng/src/test/java/core/framework/internal/validate/BeanValidatorMinMaxTest.java
@@ -8,9 +8,6 @@ import org.junit.jupiter.api.Test;
 import java.math.BigDecimal;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author neo
@@ -25,37 +22,37 @@ class BeanValidatorMinMaxTest {
 
     @Test
     void validate() {
-        Bean bean = new Bean();
+        var bean = new Bean();
         bean.num1 = 0;
         bean.num2 = 11;
 
-        ValidationErrors errors = new ValidationErrors();
+        var errors = new ValidationErrors();
         validator.validate(bean, errors, false);
 
-        assertTrue(errors.hasError());
-        assertEquals(2, errors.errors.size());
-        assertThat(errors.errors.get("num1")).contains("num1");
-        assertThat(errors.errors.get("num2")).contains("num2");
+        assertThat(errors.hasError()).isTrue();
+        assertThat(errors.errors).hasSize(2);
+        assertThat(errors.errors.get("num1")).isEqualTo("num1 must not be less than 1.0, value=0");
+        assertThat(errors.errors.get("num2")).isEqualTo("num2 must not be greater than 10.0, value=11");
     }
 
     @Test
     void validateWithoutError() {
-        Bean bean = new Bean();
+        var bean = new Bean();
         bean.num1 = 1;
         bean.num2 = 10;
         bean.num3 = new BigDecimal(6);
 
-        ValidationErrors errors = new ValidationErrors();
+        var errors = new ValidationErrors();
         validator.validate(bean, errors, false);
 
-        assertFalse(errors.hasError());
+        assertThat(errors.hasError()).isFalse();
     }
 
     static class Bean {
-        @Min(value = 1, message = "num1 must not be less than 1")
+        @Min(value = 1, message = "num1 must not be less than {min}, value={value}")
         public Integer num1;
 
-        @Max(value = 10, message = "num2 must not be greater than 10")
+        @Max(value = 10, message = "num2 must not be greater than {max}, value={value}")
         public Integer num2;
 
         @Min(5)