+++ core-ng/src/test/resources/validator-test/validator-not-blank.java
@@ -1,13 +1,13 @@
 public class BeanValidatorNotBlankTest$Bean$Validator implements core.framework.internal.validate.BeanValidator {
     private void validateBean0(core.framework.internal.validate.BeanValidatorNotBlankTest.Bean bean, core.framework.internal.validate.ValidationErrors errors, boolean partial) {
         if (bean.stringField1 == null) {
-            if (!partial) errors.add("stringField1", "field must not be null");
+            if (!partial) errors.add("stringField1", "field must not be null", null);
         } else {
-            if (bean.stringField1.isBlank()) errors.add("stringField1", "stringField1 must not be blank");
+            if (bean.stringField1.isBlank()) errors.add("stringField1", "stringField1 must not be blank", null);
         }
         if (bean.stringField2 == null) {
         } else {
-            if (bean.stringField2.isBlank()) errors.add("stringField2", "stringField2 must not be blank");
+            if (bean.stringField2.isBlank()) errors.add("stringField2", "stringField2 must not be blank", null);
         }
     }